Add DKIM Location
Just went through this with a client. Having a direct link to set up DKIM is helpful, most already have their site verified but I can create a separate pull request for that.

Question: would it be better to link to a help article such as https://support.google.com/a/answer/174126 or do you think straight to the admin console is better?